Moved SearchLayout into Search
The only purpose SearchLayout seems to be serving is to be able to consume the googleLoaded state from the GoogleMapsProvider and avoid rendering the entire layout unless googleLoaded is true. That's kind of a ham fisted approach, and I don't think I actually intended to keep it that way. Now that we can get googleLoaded from the context, it seems much better to have each individual component use it to avoid using the google maps API until the library has loaded. This will be much better for first paint since we no longer have to wait for google to be able to render anything.