Added Redux and search stores
This is essentially merging in changes from a different search project that used Next & Redux Toolkit as a simple proof of concept. The files were added then wired up with the existing project in a basic way to make sure it's working. Also added a bunch of library files, some of which the redux parts rely on, and others that may be needed in the future.