Stop passing googleMap instance for getDetails request
Redux complains whenever complex objects, such as a Google Map instance are used with it since it works best with plain objects that are serializable. I thought we needed the Map instance to create the PlacesService instance that we use to call getDetails() but apparently you can also just pass it a HTMLDivElement instead. Passing in a dummy div that's just created and never attached to the DOM appears to work just fine as an alternative, which really begs the question why you are forced to pass anything as an argument to begin with. In any case, it solves the problem of having to pass the map around inside of Redux and allows are code to be much simpler.