Fixed bug with listing search not being performed
If the geoLayerCoordinates did not change then no idle event was being triggered. Since we were relying on that exclusively the listings would be removed but no new search would be performed. Changed the initiateListingSearch() action to account for this and just perform a search immediately when this is the case.