Got search update working for user adjusting map
Updating listing results when user either drags or zooms the map. Had to fix selectCenterLatLonParams because I was getting the lat/lng from the wrong place. We need to get the center of the place we just geocoded, not the center of the map. Those two things are not the same at all.