Removed previousBounds check in bounds effect
This fixes a bug where the boundary & results were not being moved back to the center of the map when the user initiated a new search by using the search field. The bounds are the technically identical when the component gets the update, but the fact that they have been updated indicates that  we need to call fitBounds() on the map in order to make sure the boundary is placed center if the map has been dragged since the last search was initiated. I can't think of a better way to do this at the moment although there's probably a method that's more efficient to accomplish this.