Fixed bug with partial listing results being returned
After dragging the results inside the boundary outside the map viewport, results that were returned would not include anything outside the viewport, but then initiating a new search in the search field would keep this partial result from the previous search. Getting rid of this condition that only dispatches setListingSearchPending if the geolayerCoordinates have changed was causing this bug. We were getting the previous bounds that cut off the results in the request to the service because we were searching before the map had time to arrive at it's new destination. Now it will wait to search until the map idle event is triggered, and it sees that the listingSearchPending flag has been set ,instead.