Created getNextPageOfListingResults action for pagination
We want to reset the page back to 0 if we update params because the amount of pages might change, so use resetStartIndex() in searchWithUpdatedFilters. We definitely don't want to reset the page if the user is clicking to the next page so we're using the new getNextPageOfListingResults action instead of searchWithUpdatedFilters for this.