Switched to using getting geocoder & boundary data from listing service call
We were previously geocoding on the front end and getting the geolayer boundary in a separate call, then making the actual call to get the listings. The service actually support all that within one request, so I changed it to have the service take care of all this. The service seems to have been originally designed to work this way since it always returns the boundary anyway. The downside is that the map does not begin moving immediately while we are waiting for the listings to be returned as it did before. That could be potentially perceived as the app being "slower" but I compared this with other sites and in practice it seems fine. It will likely seem better once we have a loading indicator in place so the user understands what's happening. The upside is that it greatly simplifies the code and fixes some bugs I was struggling with when using the other approach.