Created selectParamsForBoundsSearch()
The code in modifyParam that changes geotype, center_lat & center_lon based on boundaryActive seemed like it should all be put together instead of existing as separate function. Decided to create new selectParamsForBoundsSearch() function to get params when the boundary is not active in order to make things more clear. The doGeospatialSearch() action is now responsible for the logic of checking if the boundary is active and getting the correct params.