Replaced Listing Marker popup with hover animation
Removed the ListingMarker Popup OverlayView with an animation that morphs the normal ListingMarkerContent view into a popup in place. The result is much slicker and gets rid of all the complex Google-API-specific code that was required to make the old Popup. Now we're just using HTML/CSS/JS with React instead. Also renamed the popupListing state to selectedListing and changed it to just require the listingid. We already have the whole listing in state so the the ID is really all that's needed to identify it.