Next iteration of URL handling
Refactored quite a bit of code for url handling. This time paying more careful attention to avoiding unnecessary renders. It's easy to get into a situation where things render in a loop if we're not careful. Renamed things and updated some comments to try and make it more clear what's going on in the code.