Fixed error caused by google not being loaded
In some circumstances the google maps api script tag has not finished downloading, which will cause an error when the <ListingMap> component tries to call `useAppSelector(selectGeoLayerBounds)`. Returning null in the selectGeoLayerBounds selector in order to avoid calling getGeoLayerBounds, which depends on google being loaded, seems to be the best solution in terms of user experience (i.e., the UI loading) so went with that rather than avoiding render of <ListingMap> completely.