Refactored Header
Renamed Header to SearchHeader and changed it into a presentational component that's just used to layout the elements in the search header.Created SearchFieldContainer and moved all the logic related to connecting to the store into that container component instead. This will serve as a wrapper for SearchField that connects it to the store. This seems like a nice separation of concerns in case we want to make SearchField reusable for a different kind of search that's not connected to redux.