Fixed error where google.maps.Map not loaded
If you go directly to the search page the Google Maps API Map library will not be loaded yet, even though the promise from the js-api-loader load method has been resolved. Guessing maybe something changed in the library after updating. To fix it we're checking for both googleLoaded as well as the presence of google.maps.Map before rendering anything map related in the ListingMap component.