Switched to maps api 3.5.3 & AdvancedMarkerElement
It looks as though the AdvancedMarkerView class that was only available in beta is no longer in beta as of api version 3.5.3. I can't find any reference to the name change but AdvancedMarkerElement shows up in the exact same place in the docs as AdvancedMarkerView and it seems to work just the same. Using this as an opportunity to get out of the beta channel. Version 3.5.3 is currently the latest release as well as the only release that has AdvancedMarkerElement so changing the version to that for now.