Fixed bug with map zoom
The map was zoomed way out every time you reloaded the search page. This was happening because selectGeoLayerBounds() was returning null when the geoLayerCoordinates were present but google.maps.LatLngBounds() was not available yet. Since there was nothing tracking google being loaded as a dependency the selector would not run again once google loaded and return the bounds. The fix was to do with useMemo inside the component instead of using a memoized selector with createSelector, since it's very easy to track googleLoaded by putting it in the dependency array for useMemo.

I decided to use createSelector with the listingMap state itself instead. It seemed a bit more efficient but also less verbose, since we were pulling in three separate selectors for most of the data in listingMap.