Removed SearchParams type from SavedSearch model
Including the type is nice because it will validate that fields are of the correct type before saving. However, this also causes Prisma to add every single field in the response, even if the field does not actually exist in the document. I really hate this behavior, and it appears that there's no way to turn it off, so I'm getting rid of the type. Will likely validate these fields using Zod in the future instead.