Get rid of exceptions in getCurrentMapState()
I've never seen these errors actually be thrown so I don't think it's actually something we need to check. The code that uses the values from getCurrentMapState seems to expect partial values from this object to begin with so returning an incomplete object will not be a problem, even if there was ever no data available for any of them.