Got basic address/no-boundary search working
When the user searches for an address, or a location that we do not have a boundary for, the Listing Service will return a response with no boundary. In this case we're using the viewport bounds from the geocodeerResult to move the map to the correct location. Once the map is adjusted, we trigger a new search based on the current map bounds using the map idle event.