#56 Handle address searches
Added logic for new location search in order to handle situations where no boundary would be returned, primarily for address searches, but also for searches where we don't have a boundary available. The new logic is handled a the new "useSearchNewLocation" custom hook. The choice to use a custom hook, instead of handling the logic inside of the thunk, primarily boiled down to having to use the router to redirect the user in some circumstances. The useRouter hook can't be used inside of a thunk. I tried just changing the listingDetail data and checking for that change inside of a useEffect hook but it wasn't working well, and it made the logic a bit hard to follow.

* Created "useSearchNewLocation" custom hook
* Created several new actions/reducers to use with the logic inside of the new hook
* Moved all listingSearch thunks into new "listingSearchCommon" module in order to avoid circular dependencies