Fixed bug & refactored PlacePredictionText
Apparently the Google Maps Autocomplete Service will sometimes return an autocomplete prediction that does not include a "main_text_matched_substrings" attribute in its structured_formatting attribute but the StructuredFormatting type does not indicate that this field is optional. Calling flatMap on this attribute when it's undefined was causing us to get into yet another endless loop. The fix was to just return if the attribute is not present. I also put the function that gets the value inside of a useMemo hook so that reference to the substrings array will remain stable and not re-render over and over, as it was doing.