Make standalone search check for listing detail
Changed standalone search to go directly to listing detail page if a listingDetail attribute was found in new search response instead of going to map search first. This means that the standalone search must do the new location search before routing to the search page. It's inefficient because the search page does the same search again after it's redirected. However, we will be able to mitigate this in the future by caching api requests using RTKQuery.

Removed the reducer that resets the listing detail in the store for newLocationGeocodeSearch.pending because causes the listing detail to blink or disappear when initiating a search from the listing detail page. It was just kind of a janky user experience and reseting the state appeared to be doing nothing for us anyway.