Added place_id to listing service request
If the user selects an autocomplete option from the search field dropdown, we will now store the place_id from the prediction and use it for the /listing/search/geocode request instead of sending the address param. The hope is that the google geocode service will be able to do the lookup faster if we send an ID rather than an address that is has to parse.