A very expensive refactor
Refactored script that generates listing data by extracting the part that does geocoding into a separate script. Ended up spending $268 in API calls while trying to figure out a good solution! Making a separate script will move the slow, expensive part of generating the data out one file that hopefully will not need to be re-generated much so we can re-generate the listing data as much as we want to refresh dates an other things.