Refactored modals
Removed the ModalRoot component that we were using in the layout in favor of just including the necessary modals on each specific page. The solution where we were dynamically rendering the correct modal component inside ModalRoot was elegant but doesn't align well with how things now work with React/Next and React-Modal.

We were getting warnings in development about re-initializing React-Modal because of double renders, which are done when React is set to strict mode in dev. This was apparently happening due to how we conditionally render the correct modal. Since the solution doesn't work well in strict mode it will likely cause problems in the future.

We also don't want dynamic component, such as ModalRoot, inside the layout because it opts component inside the layout into dynamic rendering instead of static.

The last thing is that we can hopefully avoid shipping unnecessary javascript to the client for pages that don't need all of the modals. Most pages besides the search page only need the login modal.