Dropdown refactor (#9)
* Renamed & refactored Dropdown

It's name was not very clear in what it's purpose is.

* Created ToggleOpenButton component

Extracting button out from menus to keep things DRY across the components that use the same pattern.

* Refactored MonreMenuButton

This button deviates from the default menu button styles so much that it really seems to make more sense to give it custom styles for the menu. Trying to pass a class down to the menu component and all that was way too complicated.

* Refactored ModalFooter

Renamed it so we can use it in other places without confusion.

* #31 - Add clear all filters button on desktop

* Refactored ContainedButton

* Refactored all the buttons

* Made menu buttons more accessible

* Refactored SortMenu

* Refactored ListingResultsPagination

* Moved all button menus to ControlledMenuButton

This widget just works a lot better as a controlled component

* Removed old MenuButton in favor of controlled version

Got rid of the original MenuButton. Renamed ControlledMenuButton to MenuButton.

---------

Co-authored-by: Adam Griffin <ag@agriffin.me>