Moved BedsAndBaths into a container
Attempting to follow Redux style guide recommendation to "Prefer having more UI components subscribed to the Redux store and reading data at a more granular level" while at the same time maintaining a separation of concerns between the store and a component. The hope is to make the component easier to reuse and test while at the same time avoiding having so much UI subscribed to all kinds of different state in one monolithic component.