More fixes for events
The listing search wasn't running in some circumstances, like when a user zoomed the map with their mouse scroll wheel. I also wasn't happy with the previous solution for the search running twice since it wasn't actually clear why it worked. Added back zoomChangedProgrammatically flag in order to fix the bugs and make sure we're listening for the correct events. So far this seems to handle everything correctly.

I got rid of the createEventListeners and destroyEventListeners in favor of just putting it all inside a useEffect. It doesn't look as pretty but the other way felt way too elaborate.